Add files via upload
An LSTM neural network to predict stock prices based on historical data. The model's performance is evaluated using standard metrics, and the results are visualized for better understanding. Such a model can be extended and fine-tuned for more accurate predictions, aiding in stock market analysis and decision-making.

To know more about this read my article:
https://medium.com/@harisudhan1903/predictive-analysis-and-modelling-of-stock-prices-using-lstm-networks-085d9a71214e